% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ama_export_fig.R
\name{ama_export_fig}
\alias{ama_export_fig}
\title{Function to export a figure with AMA theme}
\usage{
ama_export_fig(figure, filename, filepath)
}
\arguments{
\item{figure}{A ggplot2 object.}

\item{filename}{A character string specifying the filename without the extension.}

\item{filepath}{A character string specifying the directory to save the file.}
}
\description{
This function applies a custom theme to a ggplot2 figure and exports it to a given path.
It exports both an archived version with the current date and a current version without a date.
}
\examples{
\dontrun{
test_plot <- ggplot(mpg, aes(x=displ, y=hwy)) + geom_point()  # Create a ggplot2 plot
filename <- "sample_plot"  # Define a filename
filepath <- tempdir()  # Define a path using a temporary directory
ama_export_fig(test_plot, filename, filepath)  # Call the ama_export_fig function
}
}
