% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ama_export_tab.R
\name{ama_export_tab}
\alias{ama_export_tab}
\title{Function to export a table with AMA style}
\usage{
ama_export_tab(table, filename, filepath, caption = NULL, size = "small")
}
\arguments{
\item{table}{A data frame or matrix.}

\item{filename}{A character string specifying the filename without the extension.}

\item{filepath}{A character string specifying the directory to save the file.}

\item{caption}{A character string specifying the caption for the table.}

\item{size}{Latex size including "tiny", or "small"}
}
\description{
This function exports the provided table in both Excel (\code{.xlsx}) and LaTeX (\code{.tex}) formats.
The table is archived with the current date in the filename for the Excel version, while
the LaTeX version is saved with just the specified filename.
}
\examples{
\dontrun{
data(mtcars)  # Load the mtcars dataset
ama_export_tab(mtcars[1:5, 1:5], "sample_table", tempdir(), "Sample Caption for mtcars")
}
}
