% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ama_labs.R
\name{ama_labs}
\alias{ama_labs}
\title{Custom Label Formatting for ggplot2: American Marketing Association Style}
\usage{
ama_labs(
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  x = NULL,
  y = NULL,
  fill = NULL,
  color = NULL,
  ...
)
}
\arguments{
\item{title}{Plot title.}

\item{subtitle}{Plot subtitle.}

\item{caption}{Plot caption.}

\item{x}{X-axis label.}

\item{y}{Y-axis label.}

\item{fill}{Fill legend title.}

\item{color}{Color legend title.}

\item{...}{Additional arguments to be passed to \code{ggplot2::labs()}.}
}
\value{
Modified labels for a ggplot2 plot.
}
\description{
This function provides custom label formatting for ggplot2 based on the guidelines set by the American Marketing Association.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) + geom_point() + 
ama_labs(title = "Sample Plot") +
ama_theme()
}
}
