% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ama_theme.R
\name{ama_theme}
\alias{ama_theme}
\title{Custom Theme for ggplot2: American Marketing Association Style}
\usage{
ama_theme(
  base_size = 16,
  base_family = "sans",
  title_size = ggplot2::rel(1.2),
  axis_title_size = ggplot2::rel(1.2),
  legend_title_size = ggplot2::rel(0.6),
  legend_text_size = ggplot2::rel(0.6),
  axis_text_size = ggplot2::rel(1),
  ...
)
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Font family. Use "sans" for Arial and "serif" for Times New Roman.}

\item{title_size}{Title font size as a relative value.}

\item{axis_title_size}{Axis title font size as a relative value.}

\item{legend_title_size}{Legend title font size as a relative value.}

\item{legend_text_size}{Legend text font size as a relative value.}

\item{axis_text_size}{Axis text font size as a relative value.}

\item{...}{Additional theme elements to be passed to \code{ggplot2::theme()}.}
}
\value{
A ggplot2 theme.
}
\description{
This function provides a custom theme for ggplot2 following the guidelines set by the American Marketing Association.
}
\examples{
\dontrun{
library(ggplot2)
# Using Arial font
ggplot(mtcars, aes(mpg, wt)) + geom_point() + ama_theme()
# Using Times New Roman font
ggplot(mtcars, aes(mpg, wt)) + geom_point() + ama_theme(base_family = "serif")
}
}
