% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_assessment.R
\name{balance_assessment}
\alias{balance_assessment}
\title{Assess balance between treated and control groups}
\usage{
balance_assessment(data, treatment_col, ...)
}
\arguments{
\item{data}{A dataframe containing the data to be assessed.}

\item{treatment_col}{The name of the column that contains the treatment indicator (0 for control, 1 for treated).}

\item{...}{Names of the dependent variables.}
}
\value{
A list with two elements: 'SUR' (results of the SUR) and 'Hotelling' (results of the Hotelling's T-squared test).
}
\description{
This function performs a balance assessment between treated and control groups using
Seemingly Unrelated Regression (SUR) and Hotelling's T-squared test.
}
\examples{
\dontrun{
set.seed(123)
data = mtcars \%>\% 
  dplyr::select(mpg, cyl, disp, hp, wt) \%>\% 
  dplyr::rowwise() \%>\% 
  dplyr::mutate(treatment = sample(c(0,1), 1, replace = TRUE)) \%>\% 
  dplyr::ungroup()

results <- balance_assessment(data, "treatment", "mpg", "cyl")
print(results$SUR)
print(results$Hotelling)
}
}
