% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_balanced_panel.R
\name{get_balanced_panel}
\alias{get_balanced_panel}
\title{Extract a Balanced Panel}
\usage{
get_balanced_panel(
  data = data,
  adoption_cohort,
  lags,
  leads,
  time_var,
  unit_id_var,
  treated_period_var,
  filter_units = TRUE
)
}
\arguments{
\item{data}{The dataset to be used.}

\item{adoption_cohort}{Numeric, the specific adoption cohort.}

\item{lags}{Numeric, the number of lags.}

\item{leads}{Numeric, the number of leads.}

\item{time_var}{String, the name of the time variable.}

\item{unit_id_var}{String, the name of the unit ID variable.}

\item{treated_period_var}{String, the name of the treated period variable.}

\item{filter_units}{Logical, whether to filter only units with data on all time periods within the specified time window. Defaults to TRUE.}
}
\value{
A data frame with the balanced panel.
}
\description{
This function extracts a balanced panel from the data for a specific adoption cohort.
It drops units with missing observations at any point within its entire specified window (leads + adoption time, adoption time + lags).
Units treated in the same period as the adoption cohort are marked as "treated," and
units with their first treatment after the leads time of the specified adoption cohort
are marked as "control."
}
\examples{
\dontrun{
get_balanced_panel(data = fixest::base_stagg,
                   adoption_cohort = 5,
                   lags = 2,
                   leads = 3,
                   time_var = "year",
                   unit_id_var = "id",
                   treated_period_var = "year_treated")
}
}
