% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_tab.R
\name{nice_tab}
\alias{nice_tab}
\title{Nice Tabulation Function}
\usage{
nice_tab(data, digit_decimal = 2)
}
\arguments{
\item{data}{A data frame.}

\item{digit_decimal}{A number of decimal places.}
}
\value{
A data frame with all numeric columns rounded to the specified number of decimal places.
}
\description{
Create a custom function that takes a data frame and a number of decimal places as input,
rounds all numeric columns in the data frame to the specified number of decimal places,
and returns the modified data frame.
}
