% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coef_par_trends.R
\name{plot_coef_par_trends}
\alias{plot_coef_par_trends}
\title{Plot Coefficients of Parallel Trends}
\usage{
plot_coef_par_trends(
  data,
  dependent_vars,
  time_var,
  unit_treatment_status,
  unit_id_var,
  plot_type = "coefplot",
  combined_plot = TRUE,
  legend_position = "bottomleft",
  legend_title = "Legend Title",
  legend_args = list(),
  plot_args = list()
)
}
\arguments{
\item{data}{Data frame containing the data to be used in the model.}

\item{dependent_vars}{Named list of dependent variables to model and their respective labels.}

\item{time_var}{Name of the time variable in the data.}

\item{unit_treatment_status}{Name of the treatment status variable.}

\item{unit_id_var}{Name of the unit identification variable.}

\item{plot_type}{Type of plot to generate. Either "coefplot" or "iplot".}

\item{combined_plot}{Logical indicating whether to combine plots for all dependent variables.}

\item{legend_position}{Position of the legend on the plot.}

\item{legend_title}{Title for the legend.}

\item{legend_args}{List of additional arguments to customize the legend.}

\item{plot_args}{List of additional arguments to customize the plot.}
}
\value{
A plot visualizing interaction coefficients.
}
\description{
This function generates coefplots or iplots based on fixest outputs,
allowing the user to visualize interaction coefficients with ease.
}
\examples{
library(fixest)
data("base_did")

# Sample call to the function:
plot_coef_par_trends(
  data = base_did,
  dependent_vars = c(y = "Outcome 1", x1 = "Outcome 2"),
  time_var = "period",
  unit_treatment_status = "treat",
  unit_id_var = "id",
  plot_type = "coefplot",
  combined_plot = TRUE,
  plot_args = list(main = "Interaction coefficients Plot"),
  legend_title = "Metrics",
  legend_position = "bottomleft"
)

plot_coef_par_trends(
  data = base_did,
  dependent_vars = c(y = "Outcome 1", x1 = "Outcome 2"),
  time_var = "period",
  unit_treatment_status = "treat",
  unit_id_var = "id",
  plot_type = "coefplot",
  combined_plot = FALSE
)
}
