% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_covariate_balance_pretrend.R
\name{plot_covariate_balance_pretrend}
\alias{plot_covariate_balance_pretrend}
\title{Plot Covariate Balance Over Pre-Treatment Period}
\usage{
plot_covariate_balance_pretrend(
  balance_data,
  y_limits = c(-1, 1),
  theme_use = causalverse::ama_theme(),
  xlab = "Time to Treatment",
  ylab = "Balance (in SD unit)",
  main_title = "Covariate Balance Over Pre-Treatment Period",
  legend_title = "Covariate",
  show_legend = TRUE,
  ...
)
}
\arguments{
\item{balance_data}{A matrix containing the covariate balance data over the pre-treatment period.}

\item{y_limits}{A numeric vector of length 2 defining the y-axis limits.}

\item{theme_use}{A ggplot2 theme. By default, it uses \code{causalverse::ama_theme()}.}

\item{xlab}{A string indicating the label for the x-axis.}

\item{ylab}{A string indicating the label for the y-axis.}

\item{main_title}{A string for the main title of the plot.}

\item{legend_title}{A string for the legend title.}

\item{show_legend}{A logical; if TRUE, the legend is displayed, otherwise, it's hidden.}

\item{...}{Additional arguments passed to the ggplot labs.}
}
\value{
A ggplot2 object.
}
\description{
This function visualizes the covariate balance over the pre-treatment period.
It's particularly designed for outputs from methods like PanelMatch.
}
\examples{
\dontrun{
  balance_data_sample <- matrix(rnorm(20), nrow = 5)
  plot_covariate_balance_pretrend(balance_data_sample)
}

}
