% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_by_treatment.R
\name{plot_density_by_treatment}
\alias{plot_density_by_treatment}
\title{Plot Density by Treatment}
\usage{
plot_density_by_treatment(
  data,
  var_map,
  treatment_var,
  show_legend = TRUE,
  theme_use = ggplot2::theme_minimal(),
  ...
)
}
\arguments{
\item{data}{A data frame containing the variables to plot and a treatment variable.}

\item{var_map}{A named list mapping the column names in the data to display names for plotting.}

\item{treatment_var}{A named vector where the name is the treatment column in the data and the value is the legend title.}

\item{show_legend}{A logical value indicating whether to show the legend. Defaults to TRUE.}

\item{theme_use}{ggplot2 theme. Defaults to \code{ggplot2::theme_minimal()}.}

\item{...}{Additional arguments to be passed to \code{geom_density}.}
}
\value{
A list of ggplot objects for each variable in \code{var_map}.
}
\description{
This function creates a list of ggplot density plots for specified variables by treatment groups.
}
\examples{
\dontrun{
data(mtcars)
data <- mtcars \%>\% 
  dplyr::select(mpg, cyl) \%>\% 
  dplyr::rowwise() \%>\% 
  dplyr::mutate(treatment = sample(c(0,1), 1, replace = TRUE)) \%>\% 
  dplyr::ungroup()

plots <- plot_density_by_treatment(
  data = data,
  var_map = list("mpg" = "Var 1",
                 "cyl" = "Var 2"),
  treatment_var = c("treatment" = "Treatment Name\nin Legend")
)
}
}
