% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_par_trends.R
\name{plot_par_trends}
\alias{plot_par_trends}
\title{Plot Parallel Trends}
\usage{
plot_par_trends(
  data,
  metrics_and_names,
  treatment_status_var,
  time_var,
  conf_level = 0.95,
  non_negative = FALSE,
  display_CI = TRUE,
  output_format = "plot",
  smoothing_method = NULL
)
}
\arguments{
\item{data}{A data frame containing the data to plot.}

\item{metrics_and_names}{A named list of metrics to plot.}

\item{treatment_status_var}{The variable indicating treatment status.}

\item{time_var}{The variable indicating time.}

\item{conf_level}{Confidence level for confidence intervals (default is 0.95).}

\item{non_negative}{Logical; if TRUE, sets negative lower confidence bounds to 0.}

\item{display_CI}{Logical; if TRUE, displays confidence intervals.}

\item{output_format}{Format of the output; "plot" returns a list of ggplots, "data.frame" returns a data frame.}

\item{smoothing_method}{Method to use for smoothing; NULL means no smoothing.}
}
\value{
A list of ggplot objects or a data frame.
}
\description{
Plots parallel trends for given metrics.
}
\examples{
\dontrun{
library(tidyverse)
data <- expand.grid(entity = 1:100, time = 1:10) \%>\%
  dplyr::arrange(entity, time) \%>\%
  dplyr::mutate(
    treatment = ifelse(entity <= 50, "Treated", "Control"),
    outcome1 = 0.5 * time + rnorm(n(), 0, 2) + ifelse(treatment == "Treated", 0, 0),
    outcome2 = 3 + 0.3 * time + rnorm(n(), 0, 1) + ifelse(treatment == "Treated", 0, 2)
  )
results <- plot_par_trends(
  data = data,
  metrics_and_names = list(outcome1 = "Outcome 1", outcome2 = "Outcome 2"),
  treatment_status_var = "treatment",
  time_var = list(time = "Time"),
  smoothing_method = "loess"
)
library(gridExtra)
gridExtra::grid.arrange(grobs = results, ncol = 1)
}

}
