% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_diagnostics.R
\name{print.panel_diagnostics}
\alias{print.panel_diagnostics}
\title{Print method for panel_diagnostics objects}
\usage{
\method{print}{panel_diagnostics}(x, ...)
}
\arguments{
\item{x}{An object of class \code{panel_diagnostics}.}

\item{...}{Ignored.}
}
\description{
Print method for panel_diagnostics objects
}
\keyword{internal}
