% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_panel_estimate.R
\name{process_panel_estimate}
\alias{process_panel_estimate}
\title{Process Panel Estimate}
\usage{
process_panel_estimate(results_selected)
}
\arguments{
\item{results_selected}{A list of results from \code{panel_estimate()}. Each element in the list should be an object
containing the results for a particular estimation method. Each object must have an \code{estimate} and a \code{std.error} attribute.}
}
\value{
A data frame with columns \code{Method}, \code{Estimate}, and \code{SE}, representing the method name, the estimate value,
and the standard error, respectively. The data frame is formatted using \code{causalverse::nice_tab()}.
}
\description{
This function processes the output from \code{panel_estimate()} for panel estimates and returns a formatted data frame.
It takes a list of results, each corresponding to a different method, and combines them into a single data frame.
The data frame includes the method name, estimate, and standard error for each method.
}
\examples{
\dontrun{
library(synthdid)
setup = synthdid::panel.matrices(synthdid::california_prop99)
results_selected = panel_estimate(setup, selected_estimators = c("did", "sc"))
results_table = process_panel_estimate(results_selected)
print(results_table)
}

}
