% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid_est.R
\name{synthdid_est}
\alias{synthdid_est}
\title{Synthetic DID Estimation Using synthdid Package}
\usage{
synthdid_est(
  data,
  adoption_cohort,
  subgroup = NULL,
  lags,
  leads,
  time_var,
  unit_id_var,
  treated_period_var,
  treat_stat_var,
  outcome_var,
  seed = 1,
  method = "synthdid"
)
}
\arguments{
\item{data}{Data frame to analyze.}

\item{adoption_cohort}{Cohort in data to use as treated.}

\item{subgroup}{(Optional) List of IDs to use as treated subgroup.}

\item{lags}{Number of lags to use pre-treatment.}

\item{leads}{Number of post-treatment periods (0 for only the treatment period).}

\item{time_var}{Name of the calendar time column.}

\item{unit_id_var}{Name of the unit ID column.}

\item{treated_period_var}{Name of the treatment time period column.}

\item{treat_stat_var}{Name of the treatment indicator column.}

\item{outcome_var}{Name of the outcome variable column.}

\item{seed}{A numeric value for setting the random seed (only for placebo SE). Default is 1.}

\item{method}{The estimation method to be used. Methods include:
\itemize{
\item 'did': Difference-in-Differences.
\item 'sc': Synthetic Control Method.
\item 'sc_ridge': Synthetic Control Method with Ridge Penalty. It adds a ridge regularization to the synthetic control method when estimating the synthetic control weights.
\item 'difp': De-meaned Synthetic Control Method, as proposed by Doudchenko and Imbens (2016) and Ferman and Pinto (2021).
\item 'difp_ridge': De-meaned Synthetic Control with Ridge Penalty. It adds a ridge regularizationd when estimating the synthetic control weights.
\item 'synthdid': Synthetic Difference-in-Differences, a method developed by Arkhangelsky et al. (2021)
Defaults to 'synthdid'.
}}
}
\value{
A list containing the estimated treatment effects, standard errors, observed and predicted outcomes, synthetic control lambda weights, and counts of treated and control units.
}
\description{
This function estimates synthetic difference-in-differences using the \code{synthdid} package.
It offers a choice among \code{synthdid_estimate}, \code{did_estimate}, and \code{sc_estimate} methods
for estimation, defaulting to \code{synthdid_estimate}. It calculates treatment effects (TEs)
for each period instead of a single TE for all treated periods.
}
\examples{
\dontrun{
  library(tidyverse)
  library(causalverse)
  library(synthdid)

  data <- get_balanced_panel(
    data = fixest::base_stagg,
    adoption_cohort = 5,
    lags = 2,
    leads = 3,
    time_var = "year",
    unit_id_var = "id",
    treated_period_var = "year_treated"
  ) |>
    dplyr::mutate(treatvar = if_else(time_to_treatment >= 0, 1, 0)) |>
    dplyr::mutate(treatvar = as.integer(if_else(year_treated > (5 + 2), 0, treatvar)))

  synthdid_est(
    data,
    adoption_cohort = 5,
    lags = 2,
    leads = 3,
    time_var = "year",
    unit_id_var = "id",
    treated_period_var = "year_treated",
    treat_stat_var = "treatvar",
    outcome_var = "y"
  )
}
}
\references{
Ferman, B., & Pinto, C. (2021). Synthetic controls with imperfect pretreatment fit.
Quantitative Economics, 12(4), 1197-1221.

Doudchenko, Nikolay, and Guido W. Imbens. 2016.
“Balancing, Regression, Difference-in-Differences and Synthetic Control Methods: A Synthesis.”
NBER Working Paper 22791.

Arkhangelsky, D., Athey, S., Hirshberg, D. A., Imbens, G. W., & Wager, S. (2021).
Synthetic difference-in-differences.
American Economic Review, 111(12), 4088-4118.
}
