% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid_est_ate.R
\name{synthdid_est_ate}
\alias{synthdid_est_ate}
\title{Estimate the full Synthdid ATEs and Standard Errors}
\usage{
synthdid_est_ate(
  data,
  adoption_cohorts,
  lags,
  leads,
  time_var,
  unit_id_var,
  treated_period_var,
  treat_stat_var,
  outcome_var,
  placebo = F,
  pooled = F,
  subgroup = NULL,
  conf_level = 0.95,
  seed = 1
)
}
\arguments{
\item{data}{A data frame in long format to be analyzed.}

\item{adoption_cohorts}{Vector of cohorts to use for adoption times.}

\item{lags}{Integer, number of lags of adoption time to analyze.}

\item{leads}{Integer, number of leads of adoption time to analyze.}

\item{time_var}{String, column name of time variables.}

\item{unit_id_var}{String, ID column of units.}

\item{treated_period_var}{String, column with adoption time of each unit.}

\item{treat_stat_var}{String, column name indicating treatment status.}

\item{outcome_var}{String, column of outcome to analyze.}

\item{placebo}{Logical, whether to run placebo analysis.}

\item{pooled}{Logical, whether to run pooled analysis of all treated units.}

\item{subgroup}{Vector, IDs for subgroup analysis.}

\item{conf_level}{Numeric, confidence level for the interval estimation (Default: 95\%).}

\item{seed}{A numeric value for setting the random seed (for placebo SE and placebo analysis). Default is 1.}
}
\value{
A list containing the following elements:
\itemize{
\item{time}{: Vector of time periods used in estimation from -lags to leads (relative to the adoption period)}
\item{TE_mean}{: Vector of ATT in each time period}
\item{SE_mean}{: Vector of Standard error of ATT each time period}
\item{TE_mean_lower}{: Vector of Lower C.I. for ATT per period}
\item{TE_mean_upper}{: Vector of Upper C.I. for ATT per period}
\item{TE_mean_w, SE_mean_w, TE_mean_w_lower, TE_mean_w_upper}{: Weighted versions of the above metrics by the number of treated units in each time period }
\item{Ntr}{: Number of treated units}
\item{Nco}{: Number of control units}
\item{TE}{: Treatment effect for each cohort in each time period}
\item{SE}{: Standard error of TE of each cohort in each time period}
\item{y_obs}{: Observed outcomes of treated units}
\item{y_pred}{: Predicted outcomes of treated units}
\item{col_names}{: Column names for TE and SE matrices (times and ATTs)}
}
}
\description{
This function estimates the Average Treatment Effects (ATEs) and their
standard errors using the Synthetic Difference-in-Differences (SynthDiD)
approach. It handles different cohorts, lags and leads, placebo, and pooled
analysis options.
}
\examples{
\dontrun{
  library(tidyverse)
  data <- fixest::base_stagg |>
    mutate(treatvar = if_else(time_to_treatment >= 0, 1, 0)) |>
    mutate(treatvar = as.integer(if_else(year_treated > (5 + 2), 0, treatvar)))

  synthdid_est_ate(
    data = data,
    adoption_cohorts = 5:7,
    lags = 2,
    leads = 2,
    time_var = "year",
    unit_id_var = "id",
    treated_period_var = "year_treated",
    treat_stat_var = "treatvar",
    pooled = F,
    outcome_var = "y"
  )
}
}
