% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid_plot_ate.R
\name{synthdid_plot_ate}
\alias{synthdid_plot_ate}
\title{Create ATE Plot Using ggplot2}
\usage{
synthdid_plot_ate(
  est,
  show_CI = TRUE,
  custom_title = "",
  xlab = "Relative Time Period",
  ylab = "ATE",
  y_intercept = 0,
  theme = causalverse::ama_theme(),
  fill_color = "lightgrey"
)
}
\arguments{
\item{est}{Estimation object from \code{synthdid_est_ate}.}

\item{show_CI}{Logical; if TRUE, shows confidence intervals on the plot.}

\item{custom_title}{String; title of the plot.}

\item{xlab}{String; label for the x-axis.}

\item{ylab}{String; label for the y-axis.}

\item{y_intercept}{Numeric; value at which a horizontal line is drawn.}

\item{theme}{ggplot theme; default is set to causalverse::ama_theme().}

\item{fill_color}{String; color used for the confidence interval shading.}
}
\value{
A ggplot object representing the ATE plot.
}
\description{
This function creates a ggplot for visualizing Average Treatment Effect (ATE)
from a given estimation object.
}
\examples{
\dontrun{
  # Load required libraries
  library(ggplot2)
  library(tidyverse)
  library(causalverse)
  
  library(tidyverse)
  data <- fixest::base_stagg |>
    dplyr::mutate(treatvar = if_else(time_to_treatment >= 0, 1, 0)) |>
    dplyr::mutate(treatvar = as.integer(if_else(year_treated > (5 + 2), 0, treatvar)))
  est <- 
    synthdid_est_ate(
      data = data,
      adoption_cohorts = 5:7,
      lags = 2,
      leads = 2,
      time_var = "year",
      unit_id_var = "id",
      treated_period_var = "year_treated",
      treat_stat_var = "treatvar",
      pooled = FALSE,
      placebo = FALSE,
      outcome_var = "y"
    )
  # Generate the plot
  synthdid_plot_ate(est, show_CI = TRUE, custom_title = "Sample ATE Plot")
}
}
