% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lee_bounds.R
\name{lee_bounds}
\alias{lee_bounds}
\title{Summarize Lee Bounds for Always-Takers}
\usage{
lee_bounds(
  df,
  d,
  m,
  y,
  cluster = NULL,
  c_at_ratio = NULL,
  units = "",
  numdraws = 1000
)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{d}{Name of the treatment variable in \code{df}.}

\item{m}{Name of the mediator variable in \code{df}.}

\item{y}{Name of the outcome variable in \code{df}.}

\item{cluster}{(Optional) The name of the cluster variable for clustered bootstrapping.}

\item{c_at_ratio}{(Optional) Specifies the ratio of E\link{Y(1,1) | C}/E\link{Y(1,1) | AT}. If this is specified, the direct effect for ATs is point-identified.}

\item{units}{A string denoting the units of the outcome variable (for labeling purposes).}

\item{numdraws}{The number of bootstrap draws for estimating the standard deviation.}
}
\value{
A data frame summarizing the computed bounds with terms, estimates, and standard errors.
}
\description{
Computes and summarizes the Lee bounds on the average direct effect for always-takers (ATs) for whom there is a direct effect of treatment (D) on the outcome (Y). This function utilizes \code{\link[MedBounds]{compute_bounds_ats}} to calculate initial bounds and applies bootstrapping to estimate the standard deviation of these estimates, providing a summary in a data frame format.
}
\examples{
\dontrun{
data(example_data)
summarized_bounds <- lee_bounds(df = example_data, d = "treatment", m = "mediator", y = "outcome")
}
}
