% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_estimate.R
\name{panel_estimate}
\alias{panel_estimate}
\title{Panel Estimate Function}
\usage{
panel_estimate(
  setup,
  selected_estimators = names(panel_estimators),
  mc_replications = 200,
  seed = 1
)
}
\arguments{
\item{setup}{A list containing matrices Y, N0, and T0 for panel data analysis.}

\item{selected_estimators}{A character vector specifying which estimators to use.
Defaults to all available estimators.}

\item{mc_replications}{The number of Monte Carlo replications for computing standard errors.
Applicable if the 'mc' estimator is used. Defaults to 200.}

\item{seed}{An integer value to set the random seed for reproducibility. Defaults to 1.}
}
\value{
A list where each element corresponds to an estimator and contains its estimate and standard error.
}
\description{
This function computes estimates and standard errors for panel data using selected estimators.
It allows the user to select specific estimators and set parameters for Monte Carlo replications and seed.
}
\examples{
\dontrun{
data('california_prop99')
setup = panel.matrices(california_prop99)
results_all = panel_estimate(setup)
results_selected = panel_estimate(setup, selected_estimators = c("did", "sc"))
summary(results_selected$did$estimate)
}

}
