% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_data.R
\name{stack_data}
\alias{stack_data}
\title{Stacked Data for Staggered DiD Analysis}
\usage{
stack_data(treated_period_var, time_var, pre_window, post_window, data)
}
\arguments{
\item{treated_period_var}{A character string indicating the column name of the treatment period variable.}

\item{time_var}{A character string indicating the column name for time.}

\item{pre_window}{An integer indicating the number of periods before the treatment to consider.}

\item{post_window}{An integer indicating the number of periods after the treatment to consider.}

\item{data}{A data frame containing the dataset to be processed.}
}
\value{
A data frame with the stacked data, augmented with relative period dummy variables,
suitable for staggered DiD analysis.
}
\description{
\code{stack_data} processes datasets used in staggered Difference-in-Differences (DiD) designs.
Staggered DiD designs arise when different units (e.g., firms, regions, countries)
get treated at different time periods. This function creates cohorts based on the provided
treatment period variable and stacks them together to create a comprehensive longitudinal format
suitable for staggered DiD analyses.
}
\details{
The function emphasizes the importance of having a control group, which should be represented by
the value 10000 in the \code{treated_period_var} column of the provided dataset. The output data will
be augmented with relative period dummy variables for ease of subsequent analysis.
}
\examples{
\dontrun{
  library(did)
  library(fixest)
  data(base_stagg)
  stacked_data <- stack_data("year_treated", "year", 5, 5, base_stagg)
  feols_result <- feols(as.formula(paste0(
    "y ~ ",
    paste(paste0("`rel_period_", c(-5:-2, 0:5), "`"), collapse = " + "),
    " | id ^ df + year ^ df"
  )), data = stacked_data)
  print(feols_result)
}
}
