% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid_est.R
\name{synthdid_est}
\alias{synthdid_est}
\title{Synthetic DID Estimation Using synthdid Package}
\usage{
synthdid_est(
  data,
  adoption_cohort,
  subgroup = NULL,
  lags,
  leads,
  time_var,
  unit_id_var,
  treated_period_var,
  treat_stat_var,
  outcome_var,
  seed = 1
)
}
\arguments{
\item{data}{Data frame to analyze.}

\item{adoption_cohort}{Cohort in data to use as treated.}

\item{subgroup}{(Optional) List of IDs to use as treated subgroup.}

\item{lags}{Number of lags to use pre-treatment.}

\item{leads}{Number of post-treatment periods (0 for only the treatment period).}

\item{time_var}{Name of the calendar time column.}

\item{unit_id_var}{Name of the unit ID column.}

\item{treated_period_var}{Name of the treatment time period column.}

\item{treat_stat_var}{Name of the treatment indicator column.}

\item{outcome_var}{Name of the outcome variable column.}

\item{seed}{A numeric value for setting the random seed (only for placebo SE). Default is 1.}
}
\value{
A list containing the estimated treatment effects, standard errors, observed and predicted outcomes, synthetic control lambda weights, and counts of treated and control units.
}
\description{
This function estimates synthetic difference-in-differences using the \code{synthdid} package.
It differs from \code{synthdid::synthdid_estimate} in that it calculates treatment effects (TEs)
for each period instead of a single TE for all treated periods.
}
\examples{
\dontrun{
  library(tidyverse)
  library(causalverse)
  library(synthdid)

  data <- get_balanced_panel(
    data = fixest::base_stagg,
    adoption_cohort = 5,
    lags = 2,
    leads = 3,
    time_var = "year",
    unit_id_var = "id",
    treated_period_var = "year_treated"
  ) |>
    dplyr::mutate(treatvar = if_else(time_to_treatment >= 0, 1, 0)) |>
    dplyr::mutate(treatvar = as.integer(if_else(year_treated > (5 + 2), 0, treatvar)))

  synthdid_est(
    data,
    adoption_cohort = 5,
    lags = 2,
    leads = 3,
    time_var = "year",
    unit_id_var = "id",
    treated_period_var = "year_treated",
    treat_stat_var = "treatvar",
    outcome_var = "y"
  )
}
}
